requirements.txt
asgiref==3.4.1
Django==3.2.6
django-ckeditor==6.1.0
django-js-asset==1.2.2
django-static-fontawesome==5.14.0.0
django-widget-tweaks==1.4.8
Pillow==8.3.1
pytz==2021.1
sqlparse==0.4.1
